Move ExternalResources to utils package
